requirements.txt
SQLAlchemy ==  2.0.25
PyJWT == 2.8.0
PyMySQL== 1.1.0
Werkzeug == 3.0.1
Python-dotenv == 1.0.1
pandas == 2.2.2